Add files via upload
Finished the Meat of the code, front-end and back-end communication is fully implemented. Only missing stuff now are small things like an exception package, custom queries, JUnit, and making the HTML pretty.

The only malfunction is the "Delete Event" function, which doesn't properly delete it. Will fix that in time